GEQZ-117 :sparkles: Feat(Calendar) : 사진 S3에 저장 및 URL 불러오기 구현

+++ src/main/java/com/example/concalendar/calendar/repository/CalendarRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CalendarRepository extends JpaRepository<Calendar, Integer> {
+public interface CalendarRepository extends JpaRepository<Calendar, Long> {
 
 }
